refactor: move VisualizationErrorInfo component outside Visualization
