+++ src/components/VisualizationErrorInfo/VisualizationErrorInfo.js
